+++ src/config.ts
@@ -1,4 +1,4 @@
-const API_SERVER_URL = import.meta.env.VITE_API_SERVER_URL
+const API_SERVER_URL = process.env.API_SERVER_URL
 
 export const CONFIG = {
   API_SERVER_URL,